feature: Add initial Packer source file and Makefile
1. Why is this change neccesary?
Because we needed a Packer source file to generate a provisioned snapshot.
Because we needed a simple way to build and deploy infrastructure.

2. How does it address the issue?
By creating a packer folder that contains source files to create a provisioned
snapshot.
By creating a Makefile with basic commands to deploy and destroy ifnrastructure.

3. What side effects does this change have?
None!