feature: Use a single source of truth for Packer and Terraform variables.
1. Why is this change neccesary?
Because We currently have 3 different files for the main 3 technologies we're
using in this project (Terraform, Ansible, Packer). We need to consolidate the
values in the Packer and Terraform files to have a single file that could be
used for the two tools.

2. How does it address the issue?
By adding default values to variable definitions for both tools,
consolidating the two files into a single file with shared values and updating
the Makefile commands.

3. What side effects does this change have?
Closes #22.